+++ pkg/tasks/container/seal.go
@@ -20,6 +20,8 @@ package container
 import (
 	"context"
 	"crypto/rand"
+	"encoding/base64"
+	"encoding/json"
 	"errors"
 	"fmt"
 
@@ -39,7 +41,7 @@ type SealTask struct {
 	SealedContainerWriter    tasks.WriterProvider
 	OutputWriter             tasks.WriterProvider
 	PeerPublicKeys           []string
-	DCKDMasterKey            *memguard.LockedBuffer
+	DCKDMasterKey            string
 	DCKDTarget               string
 	JSONOutput               bool
 	DisableContainerIdentity bool
@@ -47,6 +49,7 @@ type SealTask struct {
 }
 
 // Run the task.
+//nolint:funlen,gocyclo // to refactor
 func (t *SealTask) Run(ctx context.Context) error {
 	// Check arguments
 	if types.IsNil(t.ContainerReader) {
@@ -85,17 +88,42 @@ func (t *SealTask) Run(ctx context.Context) error {
 		ss = sealv1.New()
 	}
 
-	// Convert identities to sealing keys
-	peerPublicKeys, err := ss.PublicKeys(t.PeerPublicKeys...)
-	if err != nil {
-		return fmt.Errorf("unable to transform the peer public key to a sealing key: %w", err)
-	}
-	if len(peerPublicKeys) == 0 {
-		return errors.New("at least one public key must be provided for recovery")
+	var containerKey string
+	if !t.DisableContainerIdentity {
+		opts := []seal.GenerateOption{}
+
+		// Check container sealing master key usage
+		if t.DCKDMasterKey != "" {
+			// Process target
+			if t.DCKDTarget == "" {
+				return errors.New("target flag (string) is mandatory for key derivation")
+			}
+
+			// Decode master key
+			masterKeyRaw, errDecode := base64.RawURLEncoding.DecodeString(t.DCKDMasterKey)
+			if errDecode != nil {
+				return fmt.Errorf("unable to decode master key: %w", errDecode)
+			}
+
+			// Enable deterministic container key generation
+			opts = append(opts, seal.WithDeterministicKey(memguard.NewBufferFromBytes(masterKeyRaw), t.DCKDTarget))
+		}
+
+		// Generate container key
+		containerPublicKey, containerSecretKey, errGenerate := ss.GenerateKey(opts...)
+		if errGenerate != nil {
+			return fmt.Errorf("unable to generate container key: %w", errGenerate)
+		}
+
+		// Append to identities
+		t.PeerPublicKeys = append(t.PeerPublicKeys, containerPublicKey)
+
+		// Assign container key
+		containerKey = containerSecretKey
 	}
 
 	// Seal the container
-	sealedContainer, err := ss.Seal(rand.Reader, in, peerPublicKeys...)
+	sealedContainer, err := ss.Seal(rand.Reader, in, t.PeerPublicKeys...)
 	if err != nil {
 		return fmt.Errorf("unable to seal container: %w", err)
 	}
@@ -109,6 +137,28 @@ func (t *SealTask) Run(ctx context.Context) error {
 	// Dump to writer
 	if err = container.Dump(writer, sealedContainer); err != nil {
 		return fmt.Errorf("unable to write sealed container: %w", err)
+	}
+
+	if !t.DisableContainerIdentity {
+		// Get output writer
+		outputWriter, err := t.OutputWriter(ctx)
+		if err != nil {
+			return fmt.Errorf("unable to retrieve output writer: %w", err)
+		}
+
+		// Display as json
+		if t.JSONOutput {
+			if err := json.NewEncoder(outputWriter).Encode(map[string]interface{}{
+				"container_key": containerKey,
+			}); err != nil {
+				return fmt.Errorf("unable to display as json: %w", err)
+			}
+		} else {
+			// Display container key
+			if _, err := fmt.Fprintf(outputWriter, "Container key : %s\n", containerKey); err != nil {
+				return fmt.Errorf("unable to display result: %w", err)
+			}
+		}
 	}
 
 	// No error